+++ config/prepareDist.js
@@ -53,11 +53,18 @@ fs.copyFileSync(`${srcDir}/README.md`,  `${destDir}/README.md`);
 fs.copyFileSync(`${srcDir}/LICENSE`,  `${destDir}/LICENSE`);
 
 
-/* @apollo/client/core, @apollo/client/cache, @apollo/client/utilities */
+/*
+ * @apollo/client/core
+ * @apollo/client/cache
+ * @apollo/client/utilities
+ * @apollo/client/react/ssr
+ * @apollo/client/react/hoc
+ * @apollo/client/react/components
+ */
 
-function buildPackageJson(bundleName) {
+function buildPackageJson(bundleName, entryPoint) {
   return JSON.stringify({
-    name: `@apollo/client/${bundleName}`,
+    name: `@apollo/client/${entryPoint || bundleName}`,
     main: `${bundleName}.cjs.js`,
     module: 'index.js',
     types: 'index.d.ts',
@@ -91,21 +98,42 @@ function writeCjsIndex(bundleName, exportNames, includeNames = true) {
   ].join('\n'));
 }
 
-// Create `core`, `cache` and `utilities` bundle package.json files, storing
-// them in their associated dist directory. This helps provide a way for the
-// Apollo Client core to be used without React (via `@apollo/client/core`),
-// and AC's cache and utilities to be used by themselves
-// (`@apollo/client/cache` and `@apollo/client/utilities`), via the
-// `core.cjs.js`, `cache.cjs.js` and `utilities.cjs.js` CommonJS entry point
-// files that only include the exports needed for each bundle.
+// Create individual bundle package.json files, storing them in their
+// associated dist directory. This helps provide a way for the Apollo Client
+// core to be used without React, as well as AC's cache, utilities, SSR,
+// components, and HOC to be used by themselves, via the `core.cjs.js`,
+// `cache.cjs.js`, `utilities.cjs.js`, `ssr.cjs.js`, `components.cjs.js`, and
+// `hoc.cjs.js` CommonJS entry point files that only include the exports needed
+// for each bundle.
 
+// @apollo/client/core
 fs.writeFileSync(`${distRoot}/core/package.json`, buildPackageJson('core'));
 writeCjsIndex('core', loadExportNames('react'), false);
 
+// @apollo/client/cache
 fs.writeFileSync(`${distRoot}/cache/package.json`, buildPackageJson('cache'));
 writeCjsIndex('cache', loadExportNames('cache'));
 
+// @apollo/client/utilities
 fs.writeFileSync(
   `${distRoot}/utilities/package.json`,
   buildPackageJson('utilities')
+);
+
+// @apollo/client/react/ssr
+fs.writeFileSync(
+  `${distRoot}/react/ssr/package.json`,
+  buildPackageJson('ssr', 'react/ssr')
+);
+
+// @apollo/client/react/components
+fs.writeFileSync(
+  `${distRoot}/react/components/package.json`,
+  buildPackageJson('components', 'react/components')
+);
+
+// @apollo/client/react/hoc
+fs.writeFileSync(
+  `${distRoot}/react/hoc/package.json`,
+  buildPackageJson('hoc', 'react/hoc')
 );